refactor(sdk)!: move HTTP types to the http lib and reclaim Http* for buffered (D7)
HttpAuthConfig, HttpInvocationConfig, HttpMethod and the http helper now live in
@iii-dev/http / iii-http, re-exported deprecated from the old paths. The buffered
ApiRequest/ApiResponse are renamed HttpRequest/HttpResponse (with deprecated Api*
aliases); the 2.8 Http*=Stream* streaming aliases are removed.
