+++ Cargo.toml
@@ -31,6 +31,7 @@ homepage = "https://iii.dev"
 iii-sdk = { path = "sdk/packages/rust/iii" }
 iii-observability = { path = "sdk/packages/rust/observability", version = "0.19.0" }
 iii-queue = { path = "sdk/packages/rust/queue", version = "0.19.0" }
+iii-http = { path = "sdk/packages/rust/http", version = "0.19.0" }
 scaffolder-core = { path = "crates/scaffolder-core" }
 
 # Shared dependencies used by tooling crates
