+++ sdk/packages/node/iii/package.json
@@ -64,6 +64,7 @@
     }
   },
   "dependencies": {
+    "@iii-dev/http": "workspace:*",
     "@iii-dev/observability": "workspace:*",
     "@iii-dev/queue": "workspace:*",
     "@opentelemetry/api": "^1.9.0",
