+++ sdk/packages/node/iii/tests/api-triggers.test.ts
@@ -3,8 +3,7 @@ import * as path from 'node:path'
 import { pipeline } from 'node:stream/promises'
 import { fileURLToPath } from 'node:url'
 import { describe, expect, it } from 'vitest'
-import { http, type ApiResponse, type HttpRequest } from '../src'
-import type { HttpResponse } from '../src/types'
+import { http, type ApiResponse, type StreamRequest, type StreamResponse } from '../src'
 import { engineHttpUrl, execute, httpRequest, iii, sleep } from './utils'
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url))
@@ -14,7 +13,7 @@ describe('API Triggers', () => {
   it('should register GET endpoint', async () => {
     const fn = iii.registerFunction(
       'test.api.get',
-      async (_req: HttpRequest): Promise<ApiResponse> => ({
+      async (_req: StreamRequest): Promise<ApiResponse> => ({
         status_code: 200,
         body: { message: 'Hello from GET' },
       }),
@@ -43,7 +42,7 @@ describe('API Triggers', () => {
   it('should register POST endpoint with body', async () => {
     const fn = iii.registerFunction(
       'test.api.post',
-      async (req: HttpRequest): Promise<ApiResponse> => {
+      async (req: StreamRequest): Promise<ApiResponse> => {
         const body = (req.body as Record<string, unknown>) ?? {}
         return {
           status_code: 201,
@@ -78,7 +77,7 @@ describe('API Triggers', () => {
 
     const fn = iii.registerFunction(
       'test::api::json::raw',
-      http(async (req: HttpRequest, response: HttpResponse) => {
+      http(async (req: StreamRequest, response: StreamResponse) => {
         const rawBody = await req.request_body.readAll()
 
         response.status(200)
@@ -124,7 +123,7 @@ describe('API Triggers', () => {
   it('should handle path parameters', async () => {
     const fn = iii.registerFunction(
       'test.api.getById',
-      async (req: HttpRequest): Promise<ApiResponse> => ({
+      async (req: StreamRequest): Promise<ApiResponse> => ({
         status_code: 200,
         body: { id: req.path_params?.id },
       }),
@@ -153,7 +152,7 @@ describe('API Triggers', () => {
   it('should handle query parameters', async () => {
     const fn = iii.registerFunction(
       'test.api.search',
-      async (req: HttpRequest): Promise<ApiResponse> => {
+      async (req: StreamRequest): Promise<ApiResponse> => {
         const q = req.query_params?.q
         const limit = req.query_params?.limit
         const qVal = Array.isArray(q) ? q[0] : q
@@ -189,7 +188,7 @@ describe('API Triggers', () => {
   it('should return custom status code', async () => {
     const fn = iii.registerFunction(
       'test.api.notfound',
-      async (_req: HttpRequest): Promise<ApiResponse<404>> => ({
+      async (_req: StreamRequest): Promise<ApiResponse<404>> => ({
         status_code: 404,
         body: { error: 'Not found' },
       }),
@@ -216,10 +215,11 @@ describe('API Triggers', () => {
   })
 
   it('should honor Content-Type header when returning ApiResponse with string body', async () => {
-    const xmlBody = '<?xml version="1.0" encoding="UTF-8"?><note><to>user</to><body>hello</body></note>'
+    const xmlBody =
+      '<?xml version="1.0" encoding="UTF-8"?><note><to>user</to><body>hello</body></note>'
     const fn = iii.registerFunction(
       'test.api.xml.return',
-      async (_req: HttpRequest): Promise<ApiResponse> => ({
+      async (_req: StreamRequest): Promise<ApiResponse> => ({
         status_code: 200,
         headers: { 'Content-Type': 'text/xml' },
         body: xmlBody,
@@ -250,7 +250,7 @@ describe('API Triggers', () => {
     const originalPdf = fs.readFileSync(pdfPath)
     const fn = iii.registerFunction(
       'test.api.download.pdf',
-      http(async (_req: HttpRequest, response: HttpResponse) => {
+      http(async (_req: StreamRequest, response: StreamResponse) => {
         const fileStream = fs.createReadStream(pdfPath)
 
         response.status(200)
@@ -290,7 +290,7 @@ describe('API Triggers', () => {
 
     const fn = iii.registerFunction(
       'test.api.upload.pdf',
-      http(async (req: HttpRequest, response: HttpResponse) => {
+      http(async (req: StreamRequest, response: StreamResponse) => {
         const chunks: Buffer[] = []
 
         response.status(200)
@@ -345,7 +345,7 @@ describe('API Triggers', () => {
 
     const fn = iii.registerFunction(
       'test.api.sse',
-      http(async (_req: HttpRequest, response: HttpResponse) => {
+      http(async (_req: StreamRequest, response: StreamResponse) => {
         response.status(200)
         response.headers({
           'content-type': 'text/event-stream',
@@ -419,7 +419,7 @@ describe('API Triggers', () => {
   it('should handle application/x-www-form-urlencoded request', async () => {
     const fn = iii.registerFunction(
       'test.api.form.urlencoded',
-      http(async (req: HttpRequest, response: HttpResponse) => {
+      http(async (req: StreamRequest, response: StreamResponse) => {
         const chunks: Buffer[] = []
 
         for await (const chunk of req.request_body.stream) {
@@ -482,7 +482,7 @@ describe('API Triggers', () => {
 
     const fn = iii.registerFunction(
       'test.api.form.multipart',
-      http(async (req: HttpRequest, response: HttpResponse) => {
+      http(async (req: StreamRequest, response: StreamResponse) => {
         const chunks: Buffer[] = []
 
         for await (const chunk of req.request_body.stream) {
