+++ sdk/packages/python/iii/pyproject.toml
@@ -25,6 +25,7 @@ dependencies = [
     "opentelemetry-api>=1.25",
     "iii-observability==0.19.0",
     "iii-queue==0.19.0",
+    "iii-http==0.19.0",
 ]
 
 [project.urls]
@@ -34,6 +35,7 @@ Repository = "https://github.com/iii-hq/iii"
 [tool.uv.sources]
 iii-observability = { path = "../observability", editable = true }
 iii-queue = { path = "../queue", editable = true }
+iii-http = { path = "../http", editable = true }
 
 [project.optional-dependencies]
 dev = [
