+++ sdk/packages/rust/http/Cargo.toml
@@ -16,3 +16,5 @@ name = "iii_http"
 path = "src/lib.rs"
 
 [dependencies]
+serde = { version = "1", features = ["derive"] }
+serde_json = { version = "1", features = ["unbounded_depth"] }
