+++ sdk/packages/rust/iii/src/protocol.rs
@@ -1,59 +1,9 @@
-use std::collections::HashMap;
-
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
 use uuid::Uuid;
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "UPPERCASE")]
-pub enum HttpMethod {
-    Get,
-    Post,
-    Put,
-    Patch,
-    Delete,
-}
-
-/// Authentication configuration for HTTP-invoked functions.
-///
-/// - `Hmac` -- HMAC signature verification using a shared secret.
-/// - `Bearer` -- Bearer token authentication.
-/// - `ApiKey` -- API key sent via a custom header.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(tag = "type", rename_all = "lowercase")]
-pub enum HttpAuthConfig {
-    Hmac {
-        secret_key: String,
-    },
-    Bearer {
-        token_key: String,
-    },
-    #[serde(rename = "api_key")]
-    ApiKey {
-        header: String,
-        value_key: String,
-    },
-}
-
-/// Configuration for registering an HTTP-invoked function (Lambda, Cloudflare
-/// Workers, etc.) instead of a local handler.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct HttpInvocationConfig {
-    pub url: String,
-    #[serde(default = "default_http_method")]
-    pub method: HttpMethod,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub timeout_ms: Option<u64>,
-    #[serde(default)]
-    pub headers: HashMap<String, String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub auth: Option<HttpAuthConfig>,
-}
-
-fn default_http_method() -> HttpMethod {
-    HttpMethod::Post
-}
+pub use iii_http::{HttpAuthConfig, HttpInvocationConfig, HttpMethod};
 
 /// Routing action for [`TriggerRequest`]. Determines how the engine handles
 /// the invocation.
@@ -326,6 +276,8 @@ pub struct ErrorBody {
 
 #[cfg(test)]
 mod tests {
+    use std::collections::HashMap;
+
     use super::*;
 
     #[test]
